Introduced StorageService in place of SharedWorkspaceStorage
