+++ src/commands/newConnection/CosmosDBExecuteStep.ts
@@ -8,11 +8,8 @@ import * as l10n from '@vscode/l10n';
 import { API, getExperienceFromApi } from '../../AzureDBExperiences';
 import { parseCosmosDBConnectionString } from '../../cosmosdb/cosmosDBConnectionStrings';
 import { ext } from '../../extensionVariables';
+import { type StorageItem, StorageService } from '../../services/storageService';
 import { WorkspaceResourceType } from '../../tree/workspace-api/SharedWorkspaceResourceProvider';
-import {
-    StorageImpl,
-    type StorageItem,
-} from '../../services/StorageService';
 import { type NewConnectionWizardContext } from './NewConnectionWizardContext';
 
 export class CosmosDBExecuteStep extends AzureWizardExecuteStep<NewConnectionWizardContext> {
@@ -40,7 +37,7 @@ export class CosmosDBExecuteStep extends AzureWizardExecuteStep<NewConnectionWiz
                         secrets: [connectionString],
                     };
 
-                    await StorageImpl.push(WorkspaceResourceType.AttachedAccounts, storageItem, true);
+                    await StorageService.get().push(WorkspaceResourceType.AttachedAccounts, storageItem, true);
                 },
             );
         }
