+++ src/commands/newEmulatorConnection/ExecuteStep.ts
@@ -7,11 +7,8 @@ import { AzureWizardExecuteStep } from '@microsoft/vscode-azext-utils';
 import * as l10n from '@vscode/l10n';
 import { API } from '../../AzureDBExperiences';
 import { ext } from '../../extensionVariables';
+import { type StorageItem, StorageService } from '../../services/storageService';
 import { WorkspaceResourceType } from '../../tree/workspace-api/SharedWorkspaceResourceProvider';
-import {
-    StorageImpl,
-    type StorageItem,
-} from '../../services/StorageService';
 import { type EmulatorConfiguration } from '../../utils/emulatorConfiguration';
 import {
     NewEmulatorConnectionMode,
@@ -85,9 +82,9 @@ export class ExecuteStep extends AzureWizardExecuteStep<NewEmulatorConnectionWiz
                 };
 
                 if (experience.api === API.MongoDB) {
-                    await StorageImpl.push(WorkspaceResourceType.MongoClusters, storageItem, true);
+                    await StorageService.get().push(WorkspaceResourceType.MongoClusters, storageItem, true);
                 } else {
-                    await StorageImpl.push(WorkspaceResourceType.AttachedAccounts, storageItem, true);
+                    await StorageService.get().push(WorkspaceResourceType.AttachedAccounts, storageItem, true);
                 }
             },
         );
