+++ src/tree/workspace-view/cosmosdb/CosmosDBWorkspaceItem.ts
@@ -9,6 +9,7 @@ import * as vscode from 'vscode';
 import { API, getExperienceFromApi } from '../../../AzureDBExperiences';
 import { isEmulatorSupported } from '../../../constants';
 import { ext } from '../../../extensionVariables';
+import { type StorageItem, StorageService } from '../../../services/storageService';
 import { GraphAccountAttachedResourceItem } from '../../graph/GraphAccountAttachedResourceItem';
 import { NoSqlAccountAttachedResourceItem } from '../../nosql/NoSqlAccountAttachedResourceItem';
 import { TableAccountAttachedResourceItem } from '../../table/TableAccountAttachedResourceItem';
@@ -16,7 +17,6 @@ import { type TreeElement } from '../../TreeElement';
 import { type TreeElementWithContextValue } from '../../TreeElementWithContextValue';
 import { type PersistedAccount } from '../../v1-legacy-api/AttachedAccountsTreeItem';
 import { WorkspaceResourceType } from '../../workspace-api/SharedWorkspaceResourceProvider';
-import { StorageImpl, type StorageItem } from '../../../services/StorageService';
 import { CosmosDBAttachAccountResourceItem } from './CosmosDBAttachAccountResourceItem';
 import { type CosmosDBAttachedAccountModel } from './CosmosDBAttachedAccountModel';
 import { LocalCoreEmulatorsItem } from './LocalEmulators/LocalCoreEmulatorsItem';
@@ -33,7 +33,7 @@ export class CosmosDBWorkspaceItem implements TreeElement, TreeElementWithContex
         // TODO: remove after a few releases
         await this.pickSupportedAccounts(); // Move accounts from the old storage format to the new one
 
-        const items = await StorageImpl.getItems(this.id);
+        const items = await StorageService.get().getItems(this.id);
         const children = await this.getChildrenNoEmulatorsImpl(items);
 
         if (isEmulatorSupported) {
@@ -143,7 +143,7 @@ export class CosmosDBWorkspaceItem implements TreeElement, TreeElementWithContex
                         secrets: [connectionString],
                     };
 
-                    await StorageImpl.push(WorkspaceResourceType.AttachedAccounts, storageItem, true);
+                    await StorageService.get().push(WorkspaceResourceType.AttachedAccounts, storageItem, true);
                 }
             },
         );
@@ -201,7 +201,7 @@ export class CosmosDBWorkspaceItem implements TreeElement, TreeElementWithContex
                             secrets: [connectionString],
                         };
 
-                        await StorageImpl.push(WorkspaceResourceType.AttachedAccounts, storageItem);
+                        await StorageService.get().push(WorkspaceResourceType.AttachedAccounts, storageItem);
                         await ext.secretStorage.delete(`${serviceName}.${id}`);
 
                         return storageItem;
