+++ src/tree/workspace-view/cosmosdb/LocalEmulators/LocalCoreEmulatorsItem.ts
@@ -8,11 +8,11 @@ import * as l10n from '@vscode/l10n';
 import * as vscode from 'vscode';
 import { API, getExperienceFromApi } from '../../../../AzureDBExperiences';
 import { getThemeAgnosticIconPath } from '../../../../constants';
+import { type StorageItem, StorageService } from '../../../../services/storageService';
 import { NoSqlAccountAttachedResourceItem } from '../../../nosql/NoSqlAccountAttachedResourceItem';
 import { type TreeElement } from '../../../TreeElement';
 import { type TreeElementWithContextValue } from '../../../TreeElementWithContextValue';
 import { WorkspaceResourceType } from '../../../workspace-api/SharedWorkspaceResourceProvider';
-import { StorageImpl, type StorageItem } from '../../../../services/StorageService';
 import { type CosmosDBAttachedAccountModel } from '../CosmosDBAttachedAccountModel';
 import { NewCoreEmulatorConnectionItem } from './NewCoreEmulatorConnectionItem';
 
@@ -25,7 +25,7 @@ export class LocalCoreEmulatorsItem implements TreeElement, TreeElementWithConte
     }
 
     async getChildren(): Promise<TreeElement[]> {
-        const allItems = await StorageImpl.getItems(WorkspaceResourceType.AttachedAccounts);
+        const allItems = await StorageService.get().getItems(WorkspaceResourceType.AttachedAccounts);
 
         const children = await this.getChildrenEmulatorOnlyImpl(allItems);
