+++ src/tree/workspace-view/documentdb/LocalEmulators/LocalEmulatorsItem.ts
@@ -7,12 +7,12 @@ import * as l10n from '@vscode/l10n';
 import * as vscode from 'vscode';
 import { MongoClustersExperience } from '../../../../AzureDBExperiences';
 import { getThemeAgnosticIconPath } from '../../../../constants';
+import { StorageService } from '../../../../services/storageService';
 import { type EmulatorConfiguration } from '../../../../utils/emulatorConfiguration';
 import { type ClusterModel } from '../../../documentdb/ClusterModel';
 import { type TreeElement } from '../../../TreeElement';
 import { type TreeElementWithContextValue } from '../../../TreeElementWithContextValue';
 import { WorkspaceResourceType } from '../../../workspace-api/SharedWorkspaceResourceProvider';
-import { StorageImpl } from '../../../../services/StorageService';
 import { ClusterItem } from '../ClusterItem';
 import { NewEmulatorConnectionItem } from './NewEmulatorConnectionItem';
 
@@ -25,7 +25,7 @@ export class LocalEmulatorsItem implements TreeElement, TreeElementWithContextVa
     }
 
     async getChildren(): Promise<TreeElement[]> {
-        const allItems = await StorageImpl.getItems(WorkspaceResourceType.MongoClusters);
+        const allItems = await StorageService.get().getItems(WorkspaceResourceType.MongoClusters);
         return [
             ...allItems
                 .filter((item) => item.properties?.isEmulator) // only show emulators
